requirements.txt
absl-py==1.0.0
appdirs==1.4.4
asttokens==2.0.5
astunparse==1.6.3
audioread==2.1.9
backcall==0.2.0
black==22.1.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
executing==0.8.2
Flask==2.0.2
flatbuffers==2.0
gast==0.5.3
google-api-core==2.5.0
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-cloud-speech==2.12.0
google-pasta==0.2.0
googleapis-common-protos==1.55.0
googletrans==3.0.0
grpcio==1.44.0
grpcio-status==1.44.0
h11==0.9.0
h2==3.2.0
h5py==3.6.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==4.11.1
ipykernel==6.9.0
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
libcst==0.4.1
librosa==0.9.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numba==0.55.1
numpy==1.21.0
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.0
pooch==1.6.0
prompt-toolkit==3.0.28
proto-plus==1.20.3
protobuf==3.19.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
Pygments==2.11.2
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sniffio==1.2.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
stack-data==0.1.4
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
typing-inspect==0.7.1
typing_extensions==4.1.0
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.13.3
zipp==3.7.0
